New API for saving of datasets completed. This is a backwards compatability breaking change.
You can now save a dataset in the following ways:

 - Call your_dataset.saveToFile(filename)
 - Call your_dataset.saveToFileLike(file_like_object)

In this case filename should be pointing to a location; a
new file will be created, old ones overwritten.

file_like_object can be any file like object, as described
in http://docs.python.org/lib/bltin-file-objects.html;
e.G. a socket or a string buffer or an already opened file.
Make sure that the pointer is at the right position via
the .seek() method of the file like object.

Datasets can be loaded in a similar manner:

 - DataSet.loadFromFileLike(file_like_object)
 - DataSet.loadFromFile(filename)

At the moment, DataSets are saved in pickle format. There
is some strange behaviour when using cPickle, so
serialization of rather large datasets might take longer.