Update sequential.py
This modfiication in splitWithProportion would allow to call it  before or after a _convertToOneOfMany for datasets of type ClassificationDataSet:
Case after converting: it would raise the error "IndexError: Classes and 1-of-k representation out of sync!" due to passing the data arrays in the datasetClass() instead of just the dimensions (that sets outdimension in more than one) but having incosistant the 'class' tag of the new arrays (leftDs and rightDs).
Case before converting: it is not guarantee that testDataSet and trainDataSet contain the same number of classes that the original dataset had. So, when trying to convert to many classes each dataset, there would be an out-of-index error newtarg's target would have the dimension according on how many different classes the testDataSet or traingDataSet had.